+++ server.js
@@ -4,7 +4,8 @@ var app = express(),
     webpack = require('webpack'),
     config = require('./webpack.config'),
     webpackDevMiddleware = require('webpack-dev-middleware'),
-    webpackHotMiddleware = require('webpack-hot-middleware')
+    webpackHotMiddleware = require('webpack-hot-middleware'),
+    connection = require('./blog/mysqlForServer.js')
 
 
 config.entry.unshift("webpack-hot-middleware/client")
@@ -30,10 +31,21 @@ function headFunction(res, pathname) {
     }
 }
 
+// 获取文章目录列表
+app.get('/get_catalog/:blogId', (req, res) => {
+    let blogId = req.params.blogId
+    var sql = 'select catalog from myblog where blogId = ?'
+    connection.query(sql, [blogId], (err, results) => {
+        res.send(results[0].catalog)
+    })
+})
+
+
 
 app.get('/*', (req, res) => {
     res.send('404 !!')
 })
+