+++ changeLog.md
@@ -3,3 +3,9 @@
 `2017.07.23`
 * 加入了webpack， 支持了 *react* ，用来编辑侧边栏
 * 更新了侧边栏的导航栏目，用的webpack-dev-server，暂时做不了node后台的ajax请求
+
+`2017.07.25`
+* 加入了`webpack-dev-middleware` 和 `webpack-hot-middleware` 做热更新
+* react渲染用babel的`react-hmre`
+* **【bug】** 开发时候默认到跟路径，要根据请求路径参数访问ajax还未解决 只能开发用假参数
+